requirements.txt
streamlit>=1.41.0
ultralytics>=8.3.0
opencv-python-headless
pillow>=11.0.0
ipython